swapped places of code for graph and graph2
